AND Digital Test
Hello, thank you for the opportunity with AND Digital. I have read the job spec and feel I have the experience necessary for this position. Please find my solution for the test in the SolutionAND.js file (not the ._SolutionAND.js). I have commented on each section of the algorithm to explain my approach as simply as possible - I have also provided reasons for my approach. 

Ultimately the solution works provides the expected output!